+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## Unmerged
 - Fixed (Since 4.0.1) - Unqualified call to std::move() in some opt methods
 - Fixed (Since 1.0.0) - Compile error in Clang 16 casting from cli.toWindowsCmdline
-- x Fixed (Since 6.0.0) - Divide by zero when cli.maxWidth() set to out of range value
+- Fixed (Since 6.0.0) - Divide by zero when cli.maxWidth() set to out of range value
 - Fixed (Since 6.1.0) - vcpkg build fails for x64-windows-static
 - x Fixed (Since 4.0.1) - Assert calling cli.title() immediately after creating new option
           group via cli.command(...)